+++ General/Misc.vb
@@ -3895,9 +3895,11 @@ Public Class TextEncoding
         End If
 
         If commandLine <> "" Then
-            Return commandLine.Contains(".avs") AndAlso ((commandLine.Contains("x264") AndAlso
-                Not g.ContainsPipeTool(commandLine)) OrElse (commandLine.Contains("x265") AndAlso
-                Not g.ContainsPipeTool(commandLine)))
+            Return commandLine.Contains(".avs") AndAlso
+                ((commandLine.Contains("x264") AndAlso
+                (Not g.ContainsPipeTool(commandLine) OrElse commandLine.Contains("avs2pipemod"))) OrElse
+                (commandLine.Contains("x265") AndAlso
+                (Not g.ContainsPipeTool(commandLine) OrElse commandLine.Contains("avs2pipemod"))))
         End If
     End Function
 End Class