+++ Forms/MainForm.vb
@@ -4357,10 +4357,6 @@ Public Class MainForm
             audioExist.Field = NameOf(p.FileExistAudio)
 
             b = ui.AddBool
-            b.Text = "Re-mux AAC to M4A using MP4Box"
-            b.Field = NameOf(p.RemuxAacToM4A)
-
-            b = ui.AddBool
             b.Text = "On load use AviSynth script as audio source"
             b.Help = "Sets the AviSynth script (*.avs) as audio source file when loading a source file."
             b.Field = NameOf(p.UseScriptAsAudioSource)