+++ Changelog.md
@@ -3,6 +3,8 @@
 =========================
 
 - VPY files failed to import (stax76, [#486](/../../issues/486))
+- Warning for non spec compliant AC3 bitrate moved to main dialog
+  so it's possible to ignore the warning (stax76, JKyle)
 
 
 2.1.7.6 Beta (2021-02-03)